+++ app/service/error.go
@@ -37,21 +37,3 @@ func (e centrifugeError) E() error {
 		schema.ErrDetails, details,
 	)
 }
-
-type yafudsError struct {
-	Err error
-	Msg string
-}
-
-func (e yafudsError) E() error {
-	msg := "yafuds error"
-	if e.Msg != "" {
-		msg += ": " + e.Msg
-	}
-
-	return errors.WithFields(
-		errors.WrapOrNew(e.Err, msg),
-		schema.ErrCode, service.ErrYafuds,
-		schema.ErrHTTPCode, 500,
-	)
-}