+++ app/storage/redis.go
@@ -2,13 +2,26 @@ package storage
 
 import (
 	"context"
+	"fmt"
 
 	"github.com/gomodule/redigo/redis"
 	jsoniter "github.com/json-iterator/go"
 	"gitlab.com/proemergotech/log-go/v3"
+	"gitlab.com/proemergotech/uuid-go"
 
 	"gitlab.com/proemergotech/dliver-project-skeleton/app/schema/service"
 )
+
+// todo: remove
+//  Example for lua script
+// KEYS[1]: group:%s:dummy:%s:data
+// KEYS[2]: group:%s:uuid:%s:dummy
+// ARGV[1]: <dummy_data>
+// ARGV[2]: group:%s:dummy:%s
+var saveScript = redis.NewScript(2, `
+redis.call("SET", KEYS[1], ARGV[1])
+redis.call("SET", KEYS[2], ARGV[2])
+`)
 
 type Redis struct {
 	redisPool *redis.Pool
@@ -22,11 +35,11 @@ func NewRedis(redisPool *redis.Pool, json jsoniter.API) *Redis {
 	}
 }
 
-func (rc *Redis) Close() error {
-	return rc.redisPool.Close()
+func (r *Redis) Close() error {
+	return r.redisPool.Close()
 }
 
-func (rc *Redis) closeConn(ctx context.Context, conn redis.Conn) {
+func (r *Redis) closeConn(ctx context.Context, conn redis.Conn) {
 	if err := conn.Close(); err != nil {
 		err = service.SemanticError{Msg: "failed closing redis connection, this might result in memory leak", Err: err}.E()
 		log.Warn(ctx, err.Error(), "error", err)
@@ -34,10 +47,24 @@ func (rc *Redis) closeConn(ctx context.Context, conn redis.Conn) {
 }
 
 // todo: remove
+//  Examples for key generation
+func (r *Redis) dummyID(group string, uuid uuid.UUID) string {
+	return fmt.Sprintf("group:%s:dummy:%s", group, uuid)
+}
+
+func (r *Redis) dummyKey(dummyID string) string {
+	return fmt.Sprintf("%s:data", dummyID)
+}
+
+func (r *Redis) dummyTestKey(group string, test uuid.UUID) string {
+	return fmt.Sprintf("group:%s:test:%s:dummy", group, test)
+}
+
+// todo: remove
 //  Implementation example for get simple value
-func (rc *Redis) GetSimpleFunc(ctx context.Context, key string) (string, error) {
-	conn := rc.redisPool.Get()
-	defer rc.closeConn(ctx, conn)
+func (r *Redis) GetSimpleFunc(ctx context.Context, key string) (string, error) {
+	conn := r.redisPool.Get()
+	defer r.closeConn(ctx, conn)
 
 	value, err := redis.String(conn.Do("GET", key))
 	if err != nil {
@@ -49,47 +76,26 @@ func (rc *Redis) GetSimpleFunc(ctx context.Context, key string) (string, error) 
 
 // todo: remove
 //  Implementation example for save complex value
-func (rc *Redis) SaveComplexFunc(ctx context.Context, key string, value DummyType) error {
-	conn := rc.redisPool.Get()
-	defer rc.closeConn(ctx, conn)
+func (r *Redis) SaveDummy(ctx context.Context, dummy *service.DummyType) error {
+	conn := r.redisPool.Get()
+	defer r.closeConn(ctx, conn)
 
-	body, err := rc.json.Marshal(value)
+	body, err := r.json.Marshal(dummy)
 	if err != nil {
 		return service.SemanticError{Err: err}.E()
 	}
 
-	_, err = conn.Do("SET", key, body)
+	dummyID := r.dummyID(dummy.Group, dummy.UUID)
+	_, err = saveScript.Do(
+		conn,
+		r.dummyKey(dummyID),
+		r.dummyTestKey(dummy.Group, dummy.TestUUID),
+		body,
+		dummyID,
+	)
 	if err != nil {
 		return redisError{Err: err}.E()
 	}
 
 	return nil
-}
-
-// todo: remove
-//  Implementation example for get complex value
-func (rc *Redis) GetComplexFunc(ctx context.Context, key string) (*DummyType, error) {
-	conn := rc.redisPool.Get()
-	defer rc.closeConn(ctx, conn)
-
-	repl, err := conn.Do("GET", key)
-	if err != nil {
-		return nil, redisError{Err: err}.E()
-	}
-	if repl == nil {
-		return nil, nil
-	}
-
-	result := &DummyType{}
-	if err := rc.json.Unmarshal(repl.([]byte), result); err != nil {
-		return nil, service.SemanticError{Err: err}.E()
-	}
-
-	return result, nil
-}
-
-// todo: remove
-//  Example struct with redis tag
-type DummyType struct {
-	Test string `json:"test_dummy" redis:"dummy_test"`
 }